Всё работает) выкидываем юзера в настройки gps, если у него не включен и выводим очень обидное сообщение
