+++ app/src/main/java/io/polytech/sportable/activities/MainActivity.java
@@ -2,10 +2,13 @@ package io.polytech.sportable.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.location.LocationManager;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Toast;
 
 import io.polytech.sportable.R;
 import io.polytech.sportable.activities.mapRun.MapActivity;
@@ -44,13 +47,39 @@ public class MainActivity extends AppCompatActivity {
 
             case R.id.map_button:
                 Intent map = new Intent(MainActivity.this, MapActivity.class);
-                startActivity(map);
+                if (!isGeoEnabled()) {
+                    enableGeo();
+                } else {
+                    startActivity(map);
+                    finish();
+                }
                 break;
 
             case R.id.freerun_button:
-                Intent freerun = new Intent(MainActivity.this, FreeRunActivity.class);
-                startActivity(freerun);
+                if (!isGeoEnabled()) {
+                    enableGeo();
+                } else {
+                    Intent freerun = new Intent(MainActivity.this, FreeRunActivity.class);
+                    startActivity(freerun);
+                    finish();
+                }
                 break;
         }
+    }
+
+    public boolean isGeoEnabled() {
+        Context mContext = getApplicationContext();
+        LocationManager mLocationManager = (LocationManager) mContext.getSystemService(Context.LOCATION_SERVICE);
+        boolean mIsGPSEnabled = mLocationManager.isProviderEnabled(LocationManager.GPS_PROVIDER);
+        boolean mIsNetworkEnabled = mLocationManager.isProviderEnabled(LocationManager.NETWORK_PROVIDER);
+        boolean mIsGeoDisabled = mIsGPSEnabled && mIsNetworkEnabled;
+        return mIsGeoDisabled;
+    }
+
+    public void enableGeo() {
+        Toast.makeText(this, "Еблан зачем ты зашел в карты без геолокации? Соси хуй еп)", Toast.LENGTH_SHORT).show();
+        Intent intent = new Intent(MainActivity.this, MainActivity.class);
+        startActivity(new Intent(android.provider.Settings.ACTION_LOCATION_SOURCE_SETTINGS));
+        this.finish();
     }
 }
